implemented integration tests
