+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 csv_data
 csv_data_copy
+coverage
