+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+  "env": {
+    "isTesting": "false"
+  }
+}
