+++ server/unit.test.js
@@ -1,0 +1,5 @@
+const { app } = require('./app.js');
+
+it('Testing to see if Jest works', () => {
+  expect(1).toBe(1);
+});
